Step-4a : encapsulate FloatAuto value and ctrl points.
Force everyone to use the new getters/setters for accessing FloatAuto
values, ctrl point values and ctrl point positions. This is a prereq. for
FloatAuto being able to automatically derive the ctrl point positions and
for applying automatic smoothing (or linear connections) to the ctrl points
of the bezier curve.