Step-4b : automatic propagation of FloatAuto node pos and val changes.
Implement automatic propagation of changes incl notification of
neighbouring nodes. Ctrl point positions are now automatically set to
be at 1/3 and 2/3 of the distance between nodes. Together with the removal
of the y-factor of 2 this restores the tangent property of bezier curves:
now the connection from each node to its ctrl points is at the same time
the tangent of the cuve entering the node.
This step doesn't introduce any smothing / readjusting of tangents, it
merely provides the infrastructure for doing so.

Further, this step removes the unnecessary and dangerous dependency of
the bezier calculation from playback direction. Note the latter is not
strictly necessary for this patch. It could be considered to target this
one in a separate patch. The dependency on the playback direction could
be removed from a large fraction of the gui code. It is not necessary at
all and would resolve some issues when editing masks in sigle step mode
(this was the case at least in Cinelerra 2.0).

Further changes included in this step:

 - reworked insertion of new autos to do a real bezier interpolation,
   utilising the derivative of the bezier function to calculate intial
   tangent values for the new ctrl points

 - several bugfixes done during the next years
   - reworked the interpolate_from() and made it handle more cases
   - reworked/improved the checks guarding the interpolation
   - Bugfix: interpolate_from needs the ability to take a template auto
     Basically, this rather ugly fix is imposed by the decision to use
     only one unified interpolate_from(). (I'm still convinced that
     it would be better to have a separate fill_from_template()
     operation, as I coded it in my version of the bezier patch;
     yet Pierre Dumuid insisted that there must be only a
     single function to insert new nodes
   - Bugfix: interpolating mask autos
     The preceeding changes made interploate_from() handle all
     interpolating and inserting cases. Obviousely, mask-autos
     need to handle all corner cases as well, esp. when
     interpolating the first keyframe.... :-)