Step-5 : implement the actual tangent smoothing and readjusting.
Dependant on the tangent mode of the current node, add rules to
either retain linear connections or to get a smooth curve. Further,
implement a manual mode where the tangents on both sides of a node
are coupled together.

Note the smoothing rule is not an interpolation. Rather, it is triggered
on every value and position change to restore smoth curvature. It was
specifically designed to ensure horizontal tangents on all extremal
points of the curve. The rationale for this is to prevent the curve
to go beyond this extremal points for in-between values. Note further,
this operations are triggered after loading, adding, pasting, deleting
and dragging of auto nodes as well.

There is one special rule: in the majority of cases the tangent mode
of newly added nodes is set to automatic smoothing. Only if the next
node to the left has "automatic linear" tangent mode, a newly inserted
node will inherit this setting.

This squashed commit also contains the following fixes done
in the next years:
 - retro-fitted for handling automatic tangent after copy/insert
 - code refactoring; better name the slope variable "s"
 - improved the formula for auto-smoothing, based on
   practical experience.
My auto-smoothing rule forces the tangent to become horizontal
when approaching extremal points. Acturally, I weakened this somewhat,
so the curve behaves more like the normal interpolating case, but with
a strong damping factor when approaching extremal points. The upside
is that the produced transitions look more smooth, the downside of
this is that there now is the (rare) possibility to to get an
overshooting automation curve in spite of this smoothing formula,
but I consider it a good compromise and thus an improvement.

I've used this new formula now for 4 years in several projects and
it turned out quite favourable.