Check OpenGL server version at runtime.
According to http://www.opengl.org/sdk/docs/man/xhtml/glXCreatePbuffer.xml
"glXCreatePbuffer is available only if the GLX version is 1.3 or greater."
Note that this is about GLX version. This means, that on the client side
a version of OpenGL greater than 1.3 can be installed (and in fact must be
otherwise, Cinelerra would not even start), but on the server side the
GLX version can still be lower.

If OpenGL version is less than 1.3 don't show OpenGl selection in
preferences..playback.