Check error codes in OSS audio driver.
In several places in the OSS driver, return codes are either not
checked for error, or are checked, an error is printed, and the driver
continues blindly on with invalid state.

Performed a complete review of the driver backend and for all syscalls
and ioctl()s, both *check* the return code, and also *error out* with
consistent state if an error occurs.  This corrects numerous observed
crash-or-'destroy the speakers' bugs in the OSS backend, such as a
crash when the configured device is unavailable, or full-amplitude
noise output when device setup fails.