New plugin: Blue Banana color grading/correction filter
Blue Banana offers slider-based range and non-linear (gamma)
adjustment for red, green, blue, hue, saturation and value.  It
provides feedback both through the compositor window and realtime
color histograms embedded in each adjustment slider. It can operate
on the entire frame, select specific subset of colors, the current
masked region, or a combination of a selection and a mask. It is also
able to highlight the selected region with a marching zebra.

Commit 5129a9c43dfc6fa762527e749e6756fb7e48ad23 from
http://git.xiph.org/?p=users/xiphmont/bluebanana.git;a=summary