Make relevant methods in BC_Hash use 'const char *' instead of 'char *'.
This brings CV more in line with HV where this change
already took place, and reduces the number of warnings about writable
constant strings.  It also will reduce the amount of work needed to
synchronize the plugin API between CV and HV. The change propagate
to the class StringFile as well.