Drop obsolete atexit() override in mwindow.C
In cinelerra/mwindow.C there is a reimplementation of the standard C
library function atexit().  The hack was most likely a workaround for
the problem reported in
<URL: https://www.sourceware.org/ml/libc-alpha/2001-05/msg00098.html >.

There is no code in Cinelerra using atexit() at the moment.  Removing
the hack to ensure the code build with the --with-lots-of-warnings
configure option, which detect conflicting atexit() signature in GNU
libc and Cinelerra.