Tune warning level for C and C++ code.
The extra compiler flags added to the C++ builds (-Wformat and
-Werror=format-security) were by mistake left out from the C builds.
Add them there.

Also, ensure that to compiler flags used with configure flag
--with-lots-of-warnings include the -Werror=format-security flag to fail
the same way as the normal build if a security related format error is
detected, and use more flags also for C builds, not only C++ builds.