Correct position of return after endif in fileavi.C
Correct position of return after endif in fileavi.C

The earlier commit 1b283e2b21a9 attempted to fix a Coverity complaint
that two functions each miss a return statement. Notice that in both
cases the whole function body was structured like

  #ifdef USE_AVIFLE
     if (condition)
         return something;
     else
         return 1;
  #endif

Coverity wouldn't be worth its money if it complained for a missing
return if the #ifdef and #endif lines were not present. Therefore, it
must have complained for the case when USE_AVIFILE is not defined and
the function body becomes empty. Hence, the earlier fix should have
moved the return statement after the #endif. Do that now.

Thanks to Johannes Sixt.