Fixed swapped x-y position in update() call in recordgui.C
The fifth and sixth arguments in int BC_ListBox::update() are
xposition and yposition respectively. However, when used in
recordgui.C, the order is swapped. This patch fixes it by
reinstating the correct order() are
xposition and yposition respectively. However, when used in
recordgui.C, the order is swapped. This patch fixes it by
reinstating the correct order. Found with Coverity.