Add returns in handle_opengl when missing in plugins
The function int handle_opengl in several plugins misses a return. In
several others plugins this function is returned 0. This patch adds
the return 0 where missing (consistently with other when present).
Conversion to void is not implemented since the return value is
actually used by the only caller in playback3d.C