Change if stat. to avoid a condition that leads to div by zero
A few calls (read_buffer and read_buffer) in audioalsa.C and
audiooss.C use the return value of a few calls in audiodevice.C
as their quotient. Since the return value of such calls can be
zero, this would lead to a division by zero. This patch fixes
this restructuring the if statements so that the only returns
are those that actually lead to a meaningful result (i.e. not
zero).

Thanks to Johannes Sixt for the idea and implementation.