Change int to void when return is redundant or missing
This patch changes a further number of calls from int to void
This is applied to calls defined as int, but with missing returns.
When possible reverting from int to void is also applied to other
calls that use the similar construct but that have a return. In
this case, the return is removed (this is for channelpicher/
channeledit). For one particular case, this patch changes the
 cast of wait_for_startup from int to void, and it removes its
declaration in the headers where it’s not implemented in the
corresponding C files.