Add missing returns when int-to-void not possible
While several missing returns could be fixed by converting calls
from int to void, this is not always possible. Possible causes
are the use of the return by other calls, the use of virtual
definitions, and the use of similar calls in other classes.
To avoid regressions, inconsistency with similar calls, this
patch add returns where appropriate. Note that changes to void
have been attempted and the results fall within one of the above.