Add description of motion blur plugin in English manual
The current entry in the user manual of the motion blur plugin is
 empty. Since this effect is mainly designed for real-time applications
the new entry provided by this patch describes the plugin purpose,
and it clarifies its unfitness for rendering and non-realtime
applications (for example, for adding motion blur to live-action
video). The image in the manual is also updated.

Thanks to Joe Stewart and Igor for the insight and background info
for the plugin.