Add GreyCStoration plugin to Cinelerra-CV
GREYCStoration is a rendered effect based on its counterpart
developed for GIPM. It can be optionally used as a realtime effect
but it is important to note that as a realtime effect it is slow.

GREYCstoration is an image regularization algorithm which is able to
process a color image by locally removing small variations of pixel
intensities while preserving significant global image features, such
as edges and corners. The most direct application of image
regularization is image denoising. The plugin uses the CImg library (in
particular the files GImg.h and plugins/greycstoration.h), from:
http://cimg.sourceforge.net/greycstoration/

This patch improves that from the original website:
http://sourceforge.net/projects/greycstorationcinelerraplugin

as follows:
- Improved icon (replaces the original basic one)
- picon_png.h is now created upon compilation (as the other plugins)
- Adds a comprehensive description of the functionality of the
  plugin and its parameters to the user manual.
- Code cleanup.
  - GreyCStorationWindow base class renamed PluginWindow
  - GreyCStorationConfig::interpolate is rewritten to actually
    interpolate
  - GreyCStorationConfig::interpolate variables should be int64_t,
    not long
  - Replaced #include <libintl.h> with #include "language.h"
  - greycstoration.{c,h} renamed greycstorationplugin.{c,h}
  - plugin/greycstoration.h moved out of plugin folder, which is
    also removed.
  - added greycstoration.h to makefile.
- Bug fixes:
  - converted create_objects from int to void to fix missing return
  - Fix use after free in static void* greycstoration_thread(void *arg)
    (plugins/greycstoration/plugin/greycstoration.h) lines 448 that
    led to a program crash.