Call set_done only once when closing plugin window
When rt-plugin is used as non-realtime, it is possible that
set_done is called when plugin window is already destroyed.
set_done is asyncronous operation, there is possibility that
window object is destroyed before the done event is reaching
to already destroyed window. To avoid this, make window thread
synchronous and wait window to finish.