Fixed layout of 'Asset Info' window.
A translation doesn't work well in 'Asset Info' window due to limited space.
Resized the window from 400x660 to 410x660.
Shifted the checkbox "Automatically Fix Interlacing" and the fields to the right:
 `Asset's interlacing`, `Interlace correction`, `Reel Name`, `
`Reel Number`, `Time Code Start`.
Resized the fields:
`Asset's interlacing`, `Interlace correction`, `Reel Name`, `Reel Number`.
to better fit on the window and to increase space between pulldown_arrows
of the fields and the right edge of the window.