Ignore mouse clicks on disabled popup list box button
The asset edit dialog (AssetEditWindow) contains popup list boxes to choose
interlacing options. At all times, one of the two controls is disabled.
When the user presses the mouse button over a disabled arrow button of the
list box, then moves outside the button, then releases the mouse button, a
crash occurs because the button-release handler is not prepared to for the
situation when the listbox was disabled.

Instead of checking for the disabled listbox in the button-release handler,
do not even begin any operation in button-press handler when the listbox
is disabled.