Fix narrowing error detected by GCC 10.2.0
GCC 10.2.0 reports the following error:

In file included from mwindow.h:81,
 from playbackengine.C:30:
playbackengine.C: In member function
'void PlaybackEngine::perform_change()':
transportque.inc:50:31: error: narrowing conversion of
'4294967295' from 'unsigned int' to 'int'
 50 | #define CHANGE_ALL 0xffffffff
 | ^~~~~~~~~~
playbackengine.C:180:8: note: in expansion of macro 'CHANGE_ALL'
 180 | case CHANGE_ALL:
 | ^~~~~~~~~~

The value of CHANGE_ALL does not have to be so huge.
Reduce it so that it is identified as a signed value.

Thanks to Ingo Blechschmidt https://github.com/iblech
for reporting and testing.
The fix was proposed by Einar Rünkaru.
The name and description for the patch was proposed by Johannes Sixt.